feat(uSimpleRobot): Updated to post global coordinates and allow a prefix
Updated to post global coordinates and to allow a prefix to be used. By default, the prefix is "NAV". Also updated to post `NAV_ALTITUDE` as negative depth. This is technically not correct since altitude is not the same as negative depth. However, it should be good enough for testing simple altitude behaviors.

Fixes #5